pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mindstate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.29",
    "lark>=1.2.2",
    "prompt-toolkit>=3.0.51",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "textual>=5.3.0",
    "uvicorn>=0.35.0",
]

[project.scripts]
mstate = "mindstate.__main__:_run"
mstate-api = "mindstate.api:run"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["mindstate"]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.8",
]
